Refactoring kex ecdh RFC5656
